local_bot/requirements.txt
PyQt6
pyaudio
faster-whisper
onnxruntime
httpx
python-dotenv
numpy
silero-vad


